Create functional test scripts by groups
Jenkins file is getting complicated to manage due to the large number of
introduced tests lately. As we have done with unit tests, the best way
to solve this problem is to create new scripts on npm addresing
different functional test at once.

For now, we define three groups of functional test: ws, http-get and
http-post.

Besides, this helps the developer day-to-day when checking specific
tests.